chore: base transaction + isort checks
