+++ hathor/indexes/utxo_index.py
@@ -23,8 +23,8 @@ from hathor.conf.settings import HathorSettings
 from hathor.indexes.base_index import BaseIndex
 from hathor.indexes.scope import Scope
 from hathor.transaction import BaseTransaction, Block, TxOutput
-from hathorlib.scripts import parse_address_script
 from hathor.util import sorted_merger
+from hathorlib.scripts import parse_address_script
 
 logger = get_logger()
