+++ hathor/verification/nano_header_verifier.py
@@ -46,10 +46,10 @@ from hathor.nanocontracts.utils import is_nc_public_method
 from hathor.transaction import BaseTransaction, Transaction
 from hathor.transaction.exceptions import ScriptError, TooManySigOps
 from hathor.transaction.headers.nano_header import ADDRESS_LEN_BYTES
-from hathorlib.scripts import SigopCounter, ScriptExtras
 from hathor.transaction.scripts import create_output_script, raw_script_eval
 from hathor.transaction.storage import TransactionStorage
 from hathor.verification.verification_params import VerificationParams
+from hathorlib.scripts import ScriptExtras, SigopCounter
 
 MAX_SEQNUM_DIFF_MEMPOOL = MAX_SEQNUM_JUMP_SIZE + 30
