+++ hathor/verification/transaction_verifier.py
@@ -105,8 +105,8 @@ class TransactionVerifier:
     def verify_sigops_input(self, tx: Transaction, enable_checkdatasig_count: bool = True) -> None:
         """ Count sig operations on all inputs and verify that the total sum is below the limit
         """
-        from hathorlib.scripts import SigopCounter
         from hathor.transaction.storage.exceptions import TransactionDoesNotExist
+        from hathorlib.scripts import SigopCounter
 
         counter = SigopCounter(
             max_multisig_pubkeys=self._settings.MAX_MULTISIG_PUBKEYS,
