+++ hathor/wallet/base_wallet.py
@@ -31,12 +31,12 @@ from hathor.pubsub import EventArguments, HathorEvents, PubSubManager
 from hathor.reactor import ReactorProtocol as Reactor, get_global_reactor
 from hathor.transaction import BaseTransaction, Block, TxInput, TxOutput
 from hathor.transaction.base_transaction import int_to_bytes
-from hathorlib.scripts import P2PKH, parse_address_script
 from hathor.transaction.scripts import create_output_script
 from hathor.transaction.storage import TransactionStorage
 from hathor.transaction.transaction import Transaction
 from hathor.types import AddressB58, Amount, TokenUid
 from hathor.wallet.exceptions import InputDuplicated, InsufficientFunds, PrivateKeyNotFound
+from hathorlib.scripts import P2PKH, parse_address_script
 
 logger = get_logger()
