+++ hathor/wallet/resources/nano_contracts/decode.py
@@ -18,8 +18,8 @@ import struct
 from hathor._openapi.register import register_resource
 from hathor.api_util import Resource, get_args, get_missing_params_msg, set_cors
 from hathor.transaction import Transaction
-from hathorlib.scripts import NanoContractMatchValues
 from hathor.util import json_dumpb
+from hathorlib.scripts import NanoContractMatchValues
 
 
 @register_resource
