+++ hathor/wallet/resources/nano_contracts/execute.py
@@ -21,9 +21,9 @@ from hathor._openapi.register import register_resource
 from hathor.api_util import Resource, get_missing_params_msg, render_options, set_cors
 from hathor.crypto.util import decode_address
 from hathor.transaction import Transaction, TxInput, TxOutput
-from hathorlib.scripts import P2PKH, NanoContractMatchValues
 from hathor.util import json_dumpb, json_loadb
 from hathor.wallet.exceptions import InvalidAddress
+from hathorlib.scripts import P2PKH, NanoContractMatchValues
 
 PARAMS = ['spent_tx_id', 'spent_tx_index', 'oracle_data', 'oracle_signature', 'oracle_pubkey', 'address', 'value']
