+++ hathor/wallet/resources/thin_wallet/address_balance.py
@@ -21,9 +21,9 @@ from hathor._openapi.register import register_resource
 from hathor.api_util import Resource, get_args, get_missing_params_msg, set_cors
 from hathor.conf.get_settings import get_global_settings
 from hathor.crypto.util import decode_address
-from hathorlib.scripts import parse_address_script
 from hathor.util import json_dumpb
 from hathor.wallet.exceptions import InvalidAddress
+from hathorlib.scripts import parse_address_script
 
 if TYPE_CHECKING:
     from hathor.transaction import TxOutput
