+++ hathor/wallet/util.py
@@ -21,8 +21,8 @@ from cryptography.hazmat.primitives.asymmetric import ec
 
 from hathor.conf.get_settings import get_global_settings
 from hathor.crypto.util import get_hash160, get_private_key_from_bytes
-from hathorlib.scripts import HathorScript, Opcode
 from hathor.transaction.transaction import Transaction
+from hathorlib.scripts import HathorScript, Opcode
 
 
 def generate_multisig_redeem_script(signatures_required: int, public_key_bytes: list[bytes]) -> bytes:
